for loop support in code generators
All code generators that inherit from IForCodeGenerator will now
generate for loops when a ForStatement is emitted (i.e., a for loop is
directly or indirectly generated). The C++ back-end supports this
functionality. Code generators that do not implement IForCodeGenerator
will emulate for loops with a traditional while loop construction.